sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "swarndb"
version = "0.1.0"
description = "Python SDK for SwarnDB - a high-performance vector database with virtual graph relationships"
license = {text = "BSL-1.1"}
requires-python = ">=3.9"
authors = [
    { name = "Chirotpal Das" },
]
keywords = ["vector-database", "similarity-search", "embeddings", "grpc", "numpy", "hnsw"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
async = [
    "grpcio>=1.60.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "tqdm>=4.60.0",
]

[project.urls]
Homepage = "https://github.com/swarndb/swarndb"
Documentation = "https://docs.swarndb.io"
Repository = "https://github.com/swarndb/swarndb"
Issues = "https://github.com/swarndb/swarndb/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/swarndb"]

[tool.hatch.build.targets.sdist]
include = [
    "src/swarndb/",
    "LICENSE",
    "proto/",
]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "ANN", "B", "SIM"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
